+++ UnityProject/Assets/Scripts/Items/Weapons/MagazineBehaviour.cs
@@ -41,10 +41,13 @@ public class MagazineBehaviour : NetworkBehaviour, IServerSpawn, IExaminable, IC
 	public bool isClip = false;
 	public GameObject Projectile;
 	public int ProjectilesFired = 1;
+	public bool isCell = false;
 
-	List<int> containedProjectilesFired = new List<int>();
+	[HideInInspector]
+	public List<int> containedProjectilesFired = new List<int>();
 
-	List<GameObject> containedBullets = new List<GameObject>();
+	[HideInInspector]
+	public List<GameObject> containedBullets = new List<GameObject>();
 	public AmmoType ammoType; //SET IT IN INSPECTOR
 	public int magazineSize = 20;
 
@@ -74,7 +77,7 @@ public class MagazineBehaviour : NetworkBehaviour, IServerSpawn, IExaminable, IC
 	{
 		//set to max ammo on initialization
 		clientAmmoRemains = -1;
-		if (!isClip)
+		if (!isClip && !isCell)
 		{
 			InitLists();
 		}
@@ -92,7 +95,7 @@ public class MagazineBehaviour : NetworkBehaviour, IServerSpawn, IExaminable, IC
 			containedProjectilesFired.Add(ProjectilesFired);
 		}
 	}
-	
+
 	/// <summary>
 	/// Changes size of magazine and reloads it. Be sure to call this on every client and the server if you do, or face the consequences.
 	/// Also sets the contained ammunition to full.
@@ -134,6 +137,11 @@ public class MagazineBehaviour : NetworkBehaviour, IServerSpawn, IExaminable, IC
 	/// <returns></returns>
 	public virtual void ExpendAmmo(int amount = 1)
 	{
+		if (amount < 0)
+		{
+			Logger.LogWarning("Attempted to expend a negitive amount of ammo", Category.Firearms); // dont use this method to replenish ammo
+		}
+
 		if (ClientAmmoRemains < amount)
 		{
 			Logger.LogWarning("Client ammo count is too low, cannot expend that much ammo. Make sure" +
@@ -154,7 +162,7 @@ public class MagazineBehaviour : NetworkBehaviour, IServerSpawn, IExaminable, IC
 			else
 			{
 				SyncServerAmmo(serverAmmoRemains, serverAmmoRemains - amount);
-				if (!isClip)
+				if (!isClip && !isCell)
 				{
 					for (int i = amount;i != 0;i--)
 					{
@@ -192,7 +200,7 @@ public class MagazineBehaviour : NetworkBehaviour, IServerSpawn, IExaminable, IC
 		int toTransfer = Math.Min(magazineSize - serverAmmoRemains, clip.serverAmmoRemains);
 
 		clip.ExpendAmmo(toTransfer);
-		if (!isClip)
+		if (!isClip && !isCell)
 		{
 			for (int i = toTransfer;i != 0;i--)
 			{
@@ -205,6 +213,16 @@ public class MagazineBehaviour : NetworkBehaviour, IServerSpawn, IExaminable, IC
 		ServerSetAmmoRemains(serverAmmoRemains + toTransfer);
 
 		return ("Loaded " + toTransfer + (toTransfer == 1 ? " piece" : " pieces") + " of ammunition.");
+	}
+
+	/// <summary>
+	/// method to add info to the projectile array,
+	/// should be used when ammo is being increased outside of the reload logic
+	/// </summary>
+	public void LoadProjectile(GameObject projectile,int projectilesfired)
+	{
+		containedBullets.Add(projectile);
+		containedProjectilesFired.Add(projectilesfired);
 	}
 
 	/// <summary>
@@ -249,8 +267,11 @@ public class MagazineBehaviour : NetworkBehaviour, IServerSpawn, IExaminable, IC
 
 	public void UpdateProjectile()
 	{
+		if (!isClip && !isCell)
+		{
 		ProjectilesFired = containedProjectilesFired[0];
 		Projectile = containedBullets[0];
+		}
 	}
 
 	public String Examine(Vector3 pos)