+++ UnityProject/Assets/Scripts/UI/Core/Action/UIActionManager.cs
@@ -135,7 +135,6 @@ public class UIActionManager : MonoBehaviour
 		}
 	}
 
-
 	public static void SetBackground(IActionGUI iActionGUI, Sprite sprite)
 	{
 		if (Instance.DicIActionGUI.ContainsKey(iActionGUI))
@@ -143,9 +142,30 @@ public class UIActionManager : MonoBehaviour
 			var _UIAction = Instance.DicIActionGUI[iActionGUI];
 			_UIAction.IconBackground.SetSprite(sprite);
 		}
-		else {
-			Logger.Log("iActionGUI Not present", Category.UI);
+		else
+		{
+			Logger.Log("iActionGUI not present!", Category.UI);
 		}
+	}
+
+	public static void SetCooldownLocal(IActionGUI iActionGUI, float cooldown)
+	{
+		if (Instance.DicIActionGUI.ContainsKey(iActionGUI))
+		{
+			var _UIAction = Instance.DicIActionGUI[iActionGUI];
+
+			_UIAction.CooldownOpacity.localScale = Vector3Int.one; // Enable opacity.
+			_UIAction.CooldownOpacity.LeanScaleY(0f, cooldown);
+		}
+		else
+		{
+			Logger.Log("iActionGUI not present!", Category.UI);
+		}
+	}
+
+	public static void SetCooldown(IActionGUI iActionGUI, float cooldown, GameObject recipient)
+	{
+		SetActionUIMessage.SetAction(recipient, iActionGUI, cooldown);
 	}
 
 	public static void Show(IActionGUI iActionGUI)
@@ -179,6 +199,8 @@ public class UIActionManager : MonoBehaviour
 		}
 	}
 
+	#region Events
+
 	public void OnRoundEnd()
 	{
 		foreach (var _Action in DicIActionGUI) {
@@ -202,18 +224,15 @@ public class UIActionManager : MonoBehaviour
 		CheckEvent(EVENT.LoggedOut);
 	}
 
-
 	public void RoundStarted()
 	{
 		CheckEvent(EVENT.RoundStarted);
 	}
 
-
 	public void GhostSpawned()
 	{
 		CheckEvent(EVENT.GhostSpawned);
 	}
-
 
 	public void PlayerRejoined()
 	{
@@ -240,9 +259,6 @@ public class UIActionManager : MonoBehaviour
 		}
 	}
 
-
-
-
 	private void OnEnable()
 	{
 		EventManager.AddHandler(EVENT.RoundEnded, OnRoundEnd);
@@ -268,4 +284,5 @@ public class UIActionManager : MonoBehaviour
 		EventManager.RemoveHandler(EVENT.PlayerRejoined, PlayerRejoined);
 	}
 
+	#endregion Events
 }